Modified .Xresources F7 in xterm inserts clipboard
When running Ubuntu under VirtualBox in a Mac, there's no
INSERT key so there's no way to paste the current clipboard
into the terminal.  This change makes F7 insert the contents
of CLIPBOARD into the terminal.

(There are at least 2 levels of clipboard -- PRIMARY and
CLIPBOARD.  CLIPBOARD is what X uses while PRIMARY is what
stores the values of C-x, C-w etc in xterm.  On a /normal/
keyboard, Shift-Insert inserts the CLIPBOARD contents but
the Mac, in its infinte glory and wisdom, does not provide
an Insert key.)